Add problem data model
