+++ internal/model/chat.go
@@ -1,6 +1,5 @@
 package model
 
-// 对话
 type ChatMsg struct {
 	Content string `json:"content"`
 }
