+++ internal/model/judge.go
@@ -2,6 +2,14 @@ package model
 
 type JudgeStatus uint64
 
+// 提交信息
+type Submission struct {
+	SourceCode     string `json:"source_code"`
+	Language       string `json:"language"`
+	Stdin          string `json:"stdin,omitempty"`
+	ExpectedOutput string `json:"expected_output,omitempty"`
+}
+
 // 评测结果
 type Judgement struct {
 	Stdout        string `json:"stdout,omitempty"`
