--- internal/model/translate.go
@@ -1,12 +1,0 @@
-package model
-
-// 翻译题目说明
-type TranslateInstruction struct {
-	Title       string   `json:"title,omitempty" binding:"omitempty"`
-	Description string   `json:"description,omitempty" binding:"omitempty"`
-	Input       string   `json:"input,omitempty" binding:"omitempty"`
-	Output      string   `json:"output,omitempty" binding:"omitempty"`
-	Hint        string   `json:"hint,omitempty" binding:"omitempty"`
-	Tags        []string `json:"tags,omitempty" binding:"omitempty"`
-	TargetLang  string   `json:"target_lang,omitempty" binding:"omitempty"`
-}
